requirements.txt
# Framework
fastapi
uvicorn[standard]

# Database
supabase

# AI and Data
google-generativeai
gspread
oauth2client

# Messaging
twilio

# Utilities & Environment
python-dotenv

# Testing & Linting
pytest
flake8
pytest-asyncio
pydantic-settings
python-multipart
httpx
